requirements.txt
psycopg2-binary
pymongo
pyyaml>=4.2b1